requirements.txt
tensorflow-gpu>=2.8.0
keras 
keras-applications 
scikit-learn==1.0.2
tensorflow-io
librosa